requirements/local.txt
asgiref==3.2.10
Django==3.0.8
django-cors-headers==3.4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
mysql==0.0.2
mysqlclient==2.0.1
Pillow==7.2.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1